Major bugfix: Fixed drive functions to update model with bot velocity data
four_wheel_omni_pilot now updates X/Y/AngVel data to robot data structures
located in model/robot.h and model/robot.cpp, and
called by yisicomm for transmission to physical bots.  A new
movement header (move_parameters.h) has been added for manipulating
important movement values to tweak the bot's locomotion.

Bot ID's have been altered for testing purposes.

+++ model/robot.cpp
@@ -70,7 +70,7 @@ int Robot::getLB() { return LB; }
 int Robot::getRB() { return RB; }
 int Robot::getXVel() { return xvel_cmd; }
 int Robot::getYVel() { return yvel_cmd; }
-int Robot::getAngVel() { return angvel_cmd; }
+float Robot::getAngVel() { return angvel_cmd; }
 //! @}
 
 //! @brief Get the robot's Point position on the field
